Refactor code, fix shortcut bug, and add map nav. Changes: - Refactor code to separate floor_room_class into multiple classes. - Fix shortcut bug by using modulo operator. - Add a navigation and zoom functionality for the map menu.
