+++ SFML19_RoguelikeDungeon/HeaderFiles/include.h
@@ -10,8 +10,9 @@
 
 #include "player_class.h"
 #include "enemies_class.h"
-#include "floor_room_class.h"
+#include "floor_class.h"
 #include "items_class.h"
 #include "spell_class.h"
+#include "map_class.h"
 
 #endif
